Moved transpose axes tests to test_array_manipulation, added numpy function tests for transpose
