+++ numba/tests/test_array_manipulation.py
@@ -1,12 +1,14 @@
 from __future__ import print_function
 
 from functools import partial
+from itertools import permutations
 import numba.unittest_support as unittest
 
 import numpy as np
 
 from numba.compiler import compile_isolated, Flags
 from numba import jit, types, from_dtype, errors, typeof
+from numba.errors import TypingError
 from .support import TestCase, MemoryLeakMixin, CompilationCache, tag
 
 enable_pyobj_flags = Flags()
@@ -68,6 +70,14 @@ def transpose_array(a):
 
 def numpy_transpose_array(a):
     return np.transpose(a)
+
+
+def array_transpose_axes(arr, axes):
+    return arr.transpose(axes)
+
+
+def numpy_transpose_array_axes(arr, axes):
+    return np.transpose(arr, axes)
 
 
 def squeeze_array(a):
@@ -242,6 +252,78 @@ class TestArrayManipulation(MemoryLeakMixin, TestCase):
 
         # Exceptions leak references
         self.disable_leak_check()
+
+    def test_array_transpose_axes(self):
+        pyfuncs_to_use = [numpy_transpose_array_axes, array_transpose_axes]
+
+        def run(pyfunc, arr, axes):
+            cres = self.ccache.compile(pyfunc, (typeof(arr), typeof(axes)))
+            return cres.entry_point(arr, axes)
+
+        @from_generic(pyfuncs_to_use)
+        def check(pyfunc, arr, axes):
+            expected = pyfunc(arr, axes)
+            got = run(pyfunc, arr, axes)
+            self.assertPreciseEqual(got, expected)
+            self.assertEqual(got.flags.f_contiguous,
+                             expected.flags.f_contiguous)
+            self.assertEqual(got.flags.c_contiguous,
+                             expected.flags.c_contiguous)
+
+        @from_generic(pyfuncs_to_use)
+        def check_err_axis_repeated(pyfunc, arr, axes):
+            with self.assertRaises(ValueError) as raises:
+                run(pyfunc, arr, axes)
+            self.assertEqual(str(raises.exception),
+                             "repeated axis in transpose")
+
+        @from_generic(pyfuncs_to_use)
+        def check_err_axis_oob(pyfunc, arr, axes):
+            with self.assertRaises(ValueError) as raises:
+                run(pyfunc, arr, axes)
+            self.assertEqual(str(raises.exception),
+                             "axis is out of bounds for array of given dimension")
+
+        @from_generic(pyfuncs_to_use)
+        def check_err_invalid_args(pyfunc, arr, axes):
+            with self.assertRaises((TypeError, TypingError)):
+                run(pyfunc, arr, axes)
+
+        arrs = [np.arange(24),
+                np.arange(24).reshape(4, 6),
+                np.arange(24).reshape(2, 3, 4),
+                np.arange(24).reshape(1, 2, 3, 4),
+                np.arange(64).reshape(8, 4, 2)[::3,::2,:]]
+
+        for i in range(len(arrs)):
+            for axes in permutations(tuple(range(arrs[i].ndim))):
+                ndim = len(axes)
+                neg_axes = tuple([x - ndim for x in axes])
+                check(arrs[i], axes)
+                check(arrs[i], neg_axes)
+
+        # Exceptions leak references
+        self.disable_leak_check()
+
+        check_err_invalid_args(arrs[1], "foo")
+        check_err_invalid_args(arrs[1], ("foo",))
+        check_err_invalid_args(arrs[1], 5.3)
+        check_err_invalid_args(arrs[2], (1.2, 5))
+
+        check_err_axis_repeated(arrs[1], (0,0))
+        check_err_axis_repeated(arrs[2], (2,0,0))
+        check_err_axis_repeated(arrs[3], (3,2,1,1))
+
+        check_err_axis_oob(arrs[0], (1,))
+        check_err_axis_oob(arrs[0], (-2,))
+        check_err_axis_oob(arrs[1], (0,2))
+        check_err_axis_oob(arrs[1], (-3,2))
+        check_err_axis_oob(arrs[1], (0,-3))
+        check_err_axis_oob(arrs[2], (3,1,2))
+        check_err_axis_oob(arrs[2], (-4,1,2))
+        check_err_axis_oob(arrs[3], (3,1,2,5))
+        check_err_axis_oob(arrs[3], (3,1,2,-5))
+
 
     @tag('important')
     def test_expand_dims(self):
