+++ numba/tests/test_array_methods.py
@@ -64,9 +64,6 @@ def array_T(arr):
 def array_transpose(arr):
     return arr.transpose()
 
-def array_transpose_axes(arr, axes):
-    return arr.transpose(axes)
-
 def array_copy(arr):
     return arr.copy()
 
@@ -465,68 +462,6 @@ class TestArrayMethods(MemoryLeakMixin, TestCase):
         check_arr(arr.reshape((2, 3, 4))[::2])
         arr = np.array([0]).reshape(())
         check_arr(arr)
-
-    def test_array_transpose_axes(self):
-        pyfunc = array_transpose_axes
-        def run(arr, axes):
-            cres = self.ccache.compile(pyfunc, (typeof(arr), typeof(axes)))
-            return cres.entry_point(arr, axes)
-        def check(arr, axes):
-            expected = pyfunc(arr, axes)
-            got = run(arr, axes)
-            self.assertPreciseEqual(got, expected)
-            self.assertEqual(got.flags.f_contiguous,
-                             expected.flags.f_contiguous)
-            self.assertEqual(got.flags.c_contiguous,
-                             expected.flags.c_contiguous)
-        def check_err_axis_repeated(arr, axes):
-            with self.assertRaises(ValueError) as raises:
-                run(arr, axes)
-            self.assertEqual(str(raises.exception),
-                             "repeated axis in transpose")
-        def check_err_axis_oob(arr, axes):
-            with self.assertRaises(ValueError) as raises:
-                run(arr, axes)
-            self.assertEqual(str(raises.exception),
-                             "axis is out of bounds for array of given dimension")
-        def check_err_invalid_args(arr, axes):
-            with self.assertRaises((TypeError, TypingError)):
-                run(arr, axes)
-
-        arrs = [np.arange(24),
-                np.arange(24).reshape(4, 6),
-                np.arange(24).reshape(2, 3, 4),
-                np.arange(24).reshape(1, 2, 3, 4),
-                np.arange(64).reshape(8, 4, 2)[::3,::2,:]]
-
-        for i in range(len(arrs)):
-            for axes in permutations(tuple(range(arrs[i].ndim))):
-                ndim = len(axes)
-                neg_axes = tuple([x - ndim for x in axes])
-                check(arrs[i], axes)
-                check(arrs[i], neg_axes)
-
-        # Exceptions leak references
-        self.disable_leak_check()
-
-        check_err_invalid_args(arrs[1], "foo")
-        check_err_invalid_args(arrs[1], ("foo",))
-        check_err_invalid_args(arrs[1], 5.3)
-        check_err_invalid_args(arrs[2], (1.2, 5))
-
-        check_err_axis_repeated(arrs[1], (0,0))
-        check_err_axis_repeated(arrs[2], (2,0,0))
-        check_err_axis_repeated(arrs[3], (3,2,1,1))
-
-        check_err_axis_oob(arrs[0], (1,))
-        check_err_axis_oob(arrs[0], (-2,))
-        check_err_axis_oob(arrs[1], (0,2))
-        check_err_axis_oob(arrs[1], (-3,2))
-        check_err_axis_oob(arrs[1], (0,-3))
-        check_err_axis_oob(arrs[2], (3,1,2))
-        check_err_axis_oob(arrs[2], (-4,1,2))
-        check_err_axis_oob(arrs[3], (3,1,2,5))
-        check_err_axis_oob(arrs[3], (3,1,2,-5))
 
 
     def test_array_transpose(self):
